move graph to own module, remove intermediate basenetwork, attach network algorithms
