+++ src/compas/datastructures/graph/__init__.py
@@ -1,0 +1,1 @@
+from .graph import Graph  # noqa: F401
