+++ src/compas/datastructures/graph/graph.py
