+++ src/compas/datastructures/network/__init__.py
@@ -2,8 +2,14 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from .core import *  # noqa: F401 F403
-from ._network import *  # noqa: F401 F403
+import compas
+
+from .operations import *  # noqa: F401 F403
+
+if not compas.IPY:
+    from .matrices import *  # noqa: F401 F403
+
+from .network import *  # noqa: F401 F403
 
 from .combinatorics import *  # noqa: F401 F403
 from .complementarity import *  # noqa: F401 F403
