--- src/compas/datastructures/network/core/__init__.py
@@ -1,14 +1,0 @@
-from __future__ import print_function
-from __future__ import absolute_import
-from __future__ import division
-
-from compas import IPY
-
-from .graph import Graph  # noqa: F401
-from .network import BaseNetwork  # noqa: F401
-
-from .operations import *  # noqa: F401 F403
-if not IPY:
-    from .matrices import *  # noqa: F401 F403
-
-__all__ = [name for name in dir() if not name.startswith('_')]
