+++ src/compas/datastructures/network/matrices.py
