+++ src/compas/datastructures/network/operations/__init__.py
