+++ src/compas/datastructures/network/operations/join.py
