+++ src/compas/datastructures/network/operations/split.py
